Seperate tests and add further tests.
