+++ estimagic/inference/bootstrap.py
@@ -1,5 +1,6 @@
 import pandas as pd
 
+from estimagic.inference.bootstrap_ci import _check_inputs
 from estimagic.inference.bootstrap_ci import compute_ci
 from estimagic.inference.bootstrap_estimates import get_bootstrap_estimates
 from estimagic.inference.bootstrap_samples import get_seeds
@@ -51,7 +52,6 @@ def bootstrap(
 
     results = get_results_table(df, f, estimates, ci_method, alpha, num_threads)
 
-    # should I also return the seeds/sample indices?
     if return_seeds is False:
         return results
 
@@ -78,6 +78,8 @@ def get_results_table(
 
     """
 
+    _check_inputs(data=data, ci_method=ci_method, alpha=alpha)
+
     results = pd.DataFrame(estimates.mean(axis=0), columns=["mean"])
 
     results["std"] = estimates.std(axis=0)
@@ -87,34 +89,3 @@ def get_results_table(
     results["upper_ci"] = cis["upper_ci"]
 
     return results
-
-
-def _check_inputs(data, cluster_by=None, ci_method="percentile", alpha=0.05):
-    """ Check validity of inputs.
-    Args:
-        data (pd.DataFrame): original dataset.
-        cluster_by (str): column name of variable to cluster by.
-        ci_method (str): method of choice for confidence interval computation.
-        alpha (float): significance level of choice.
-
-    """
-
-    ci_method_list = ["percentile", "bca", "bc", "t", "normal", "basic"]
-
-    if not isinstance(data, pd.DataFrame):
-        raise ValueError("Input 'data' must be DataFrame.")
-
-    elif (cluster_by is not None) and (cluster_by not in data.columns.tolist()):
-        raise ValueError(
-            "Input 'cluster_by' must be None or a column name of DataFrame."
-        )
-
-    elif ci_method not in ci_method_list:
-        raise ValueError(
-            "ci_method must be 'percentile', 'bc',"
-            " 'bca', 't', 'basic' or 'normal', '{method}'"
-            " was supplied".format(method=ci_method)
-        )
-
-    elif alpha > 1 or alpha < 0:
-        raise ValueError("Input 'alpha' must be in [0,1].")
